+++ nix/vendor-hashes.nix
@@ -1,10 +1,10 @@
 {
-  root = "sha256-j4F2nhhqXx55DtgIID6ux3MVy9pVBQmusmAiHJXTpPs=";
-  laravel9 = "sha256-XXcuEu7qs9nagf3LuJY5tcY6ySJA8K6hlLBnViPhAxg=";
-  laravel11 = "sha256-N51ABU5vx3mLeR7qQWdz4n+4Pwel0zK/gs98uVMnHlQ=";
-  laravel12 = "sha256-k5FDA6k0A/tJGezCfx6W1UzMFkYMvdrEyKwUlvN4+NM=";
-  laravel13 = "sha256-IFWZc01kn2VMaio0SYwkB+7QeoS5e5xlJaIfal15aTo=";
-  lowest = "sha256-7kb2VbzZ43XXnBqYGsA1Uj8biI2IkHMHAGw8tsiGi0U=";
-  eris = "sha256-zC0ohdWMhvp+NY0PLOORmw5YXnechp7ll2RBnovNVKU=";
-  akashi = "sha256-jUtgGvUyQX5KtcfI2iICp6wMMTXqsYuk96XemBigZXQ=";
+  root = "sha256-xlbbrA4NK+Bn7lJdNFGM/0rzv8ALcSUFB9w1FtxsHjw=";
+  laravel9 = "sha256-RuIjB1lwdRMChmWKOiJrmCdJ1m9hWVBs8faqYX5AI8w=";
+  laravel11 = "sha256-D8NrhXwE+Cufc3wn3rIEdtXu2adzXyts40dR7xQOoUs=";
+  laravel12 = "sha256-Bpqms5CcWQnv5+rlKnpDf0wG9RNfCWfhHIbSbCI8HeM=";
+  laravel13 = "sha256-a63d/ABeJQ4DV0sgD73xxCM8MyRz1ujR1pGjdTexgg8=";
+  lowest = "sha256-aJf4+b7ATaURHNhgCcqS19Qh0BpJEwyrpREE41+99qA=";
+  eris = "sha256-ayU9E8tq1KsN7H+Wa1G86BY3NkL8LiQtj4vwkKv1OPM=";
+  akashi = "sha256-xMnNT2wAyDNjbVJfvXTK6rPmnxNf6NPjadfHVv09uqo=";
 }